mirgrations from 3rd party packages won't be deleted any more (as long as they reside in a virtualenv folder called myenv)

+++ webpage/management/commands/delete_migrations.py
@@ -16,9 +16,12 @@ class Command(BaseCommand):
         for root, dirs, files in os.walk(cwd):
             for file in files:
                 if 'migrations' in os.path.join(root, file) and '00' in os.path.join(root, file):
-                    deleted_files.append((os.path.join(root, file)))
-                    os.remove(os.path.join(root, file))
-                    counter = +1
+                    if 'myenv' in os.path.join(root, file):
+                        pass
+                    else:
+                        deleted_files.append((os.path.join(root, file)))
+                        os.remove(os.path.join(root, file))
+                        counter = +1
         self.stdout.write("Following {} files have been deleted".format(counter))
         for x in deleted_files:
             self.stdout.write("Deleted: {}".format(x))
